Update patch set 13
Patch Set 13:

> Patch Set 13: Code-Review+1
> 
> The performance defaults look good to me.  I'll look again and give a +2 after I see a passing Dallas job. Thanks!

Thanks for taking a look to this one.  At the moment it has passed all the ocp vanilla pipelines, but just failed in the example-cnf for 4.8 which is currently failing and 4.6, this one will need to be reviewed but we are working on some changes for it, based in our work in NM Lab.

I'm including the last runs for OCP after the last patchset, all these are listed above, but is not that clear :)  

- 4.8 https://www.distributed-ci.io/jobs/976bec7a-e946-438b-821c-3b30a0b87c37/jobStates
- 4.7 https://www.distributed-ci.io/jobs/c494fa68-3f8f-44dd-9a96-2442ef3b8b59/jobStates
- 4.6 https://www.distributed-ci.io/jobs/0305c3f4-92eb-4850-8eb5-066b78e8ddaa/jobStates
- 4.5 https://www.distributed-ci.io/jobs/a7e902c0-cec1-4bc7-b272-6ca4fd510d3b/jobStates

Patch-set: 13
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>