Update patch set 6
Patch Set 6:

Finish condition on task "Wait for upgrade to start" seems not to be well defined, causing the step "Monitor de upgrade until complete" to prematurely detect the end of the upgrade.

Changing task "Wait for upgrade to start" to wait for condition "Progressing" message to change to "Working towards {{ version }}".

Patch-set: 6