Update patch set 21
Patch Set 21:

Thanks for the help, Charles and Fred.

Here's the thing now. The DCI job runs the pipeline and, when it gets to the "monitor the upgrade" step, the process fails and ansible detects it, causing the job to fail, as expected due to this CR.

Can this be considered a successful test?

Patch-set: 21