Update patch set 8
Patch Set 8:

I need another set of eyes on this.

This works in my home lab with 4.7..  It fails in Dallas if dci_prepare_cnf=True, but if I disable dci_prepare_cnf it passes as you can see below.

When I look at the failure job it never finishes initializing the csv for cnv.

Patch-set: 8