Update patch set 1
Patch Set 1:

I remember that, in this CR from the app-agent (https://softwarefactory-project.io/r/c/dci-openshift-app-agent/+/22368), a similar change was done, but dci_teardown_on_success variable was set to true by default (here, it is set to false), and the dci_teardown_on_failure variable was used in plays/error.yml (and here, it is not used, and set to true instead of false). Just commenting this in case you want to follow the same approach than in the app-agent.

Patch-set: 1
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>