Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I remember that, in this CR from the app-agent (https://softwarefactory-project.io/r/c/dci-openshift-app-agent/+/22368), a similar change was done, but dci_teardown_on_success variable was set to true by default (here, it is set to false), and the dci_teardown_on_failure variable was used in plays/error.yml (and here, it is not used, and set to true instead of false). Just commenting this in case you want to follow the same approach than in the app-agent.

Thanks Ramón, in fact I was thinking the same after submitting this.

I think what we want, as you pointed out, is to match the same behavior as in d-o-a-a. In the case of partners if things need to remain after d-o-a successfully runs then the variable needs to be passed as part of a pipeline to not teardown things.

Patch-set: 1