Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think you should try with a regular vanilla job for this, you're doing f5-bigip

Oh, the reason of the last job was to prepare it for nokia-chf, in that job I changed the value of dci_teardown_on_success to false as I didn't want it to destroy the environment, so d-o-a-a with nokia could run on top if it.

(I did: DCI_QUEUE_RESOURCE=cluster5 check https://github.com/dci-labs/dallas-pipelines/pull/106 ocp-4.5-vcp openshift-vcp:ansible_extravars=dci_teardown_on_success:false,)

The validation of this change has been already done with the libvirt runs.

Patch-set: 2