Update patch set 14
Patch Set 14:

(1 comment)

The current merge conflict was due to this change: 22454 I approve that one as a temporary fix to get VCP in 4.7. This change is a more permanent fix, and will take precedence once we are ready to merge this one.

Patch-set: 14
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>