Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If this setting is only required for SPK, wouldn't it be better to provide a performance profile for it? The performance_definition var allows to specify a file to use.
> https://github.com/redhat-cip/dci-openshift-agent/blob/master/roles/operator-performance-profile/README.md?plain=1#L12

We could do that, but should we?
Hugepages is currently set to 1G with d.o.a, on the other hand it is set to 2Mi on cluster1 and cluster4. Why do we need this value so high? I think we should stick to the same value used by the other team.

Patch-set: 1
Reviewer: Gerrit User 673 <673@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>