Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > If this setting is only required for SPK, wouldn't it be better to provide a performance profile for it? The performance_definition var allows to specify a file to use.
> > https://github.com/redhat-cip/dci-openshift-agent/blob/master/roles/operator-performance-profile/README.md?plain=1#L12
> 
> We could do that, but should we?
> Hugepages is currently set to 1G with d.o.a, on the other hand it is set to 2Mi on cluster1 and cluster4. Why do we need this value so high? I think we should stick to the same value used by the other team.

As discussed yesterday in the daily, it seems that the requirement of 1G is due to example-cnf. So, should we think on using a performance profile for SPK? Or, as 2Mi is also the default value on cluster1 and cluster4, should this profile be done for example-cnf instead?

Patch-set: 1