Specify kubeconfig env variable for operators task
Currently deploy-operators tasks use the kubeconfig env variable
from the .bashrc file to interact with the OCP API, this variable
is injected by ansible-ipi roles, however SNO inventory uses
ansible_connection=local to perform actions on the same host and
is not taking that variable into account. In preparation of SNO
integration this includes the kubeconfig variable explicitly in
the tasks from this play.

Change-Id: Ic8f856dc0747a3faf74c43794c129f720905d11a