Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3:
> 
> (3 comments)
> 
> @Jorge @Tony
> 
> I failed to pass this in the commit message or code comments:
> 
> The current components don't have their "url" field set so I kept the "build" and "version" etc. vars & logic to be backwards-compatible with the current components.
> 
> This can be removed later.
> 
> WDYT ?

I think is OK to remove things later if the data.url field has only being populated lately. I guess I could go and search the components with the new field in place, but if you have that info at hand it would be nice to add a comment with the cut versions so in the future we can remove this code as it's a bit complicated.

I'm open to suggestions or other ideas, even to move this discussion off this change. I wanted to  share this thought as I really would like to make it easy to find out when it would be a good time to remove the additional vars.

Patch-set: 3
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1