Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> I have one more comment (maybe other people can correct me if I am wrong). You can also think on defining the new variables you are creating (max_image_age and cleanup_old_images) in a specific file. I think there are two ways:
> - Using the dci-openshift-agent/group_vars/all file (but maybe it is a generic file).
> - Moving your new code to a role, under dci-openshift-agent/roles/<your_role>/tasks/main.yml, and define the new variables in dci-openshift-agent/roles/<your_role>/defaults/main.yml. You can check the roles directory to see examples about how to do this.

Hi @Ramon, thank you for bringing that up. We discussed where to place these variables with @Charles and decided in favor of inventories simply because the idea to clean up the cache images seems to be pretty high-level one and not a role-specific. That was quite a fast decision though, and we could re-discuss during today daily.

Patch-set: 15