Update patch set 2
Patch Set 2:

Hi Tatiana! I'd just like to comment something, also because I am not sure about this and if somebody could confirm it, it would be great.
I think you can directly execute the check command in your PR in the baremetal-deploy repository, instead of creating this empty change to run the pipeline. But again, I am not 100% sure about this.

Patch-set: 2
CC: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>