Update patch set 36
Patch Set 36:

Since we only have one cluster at the moment, would it be possible to test this in libvirt? and once validated then run it in dallas? I'm trying to see if there's any way to reduce the use of tests in dallas and at the same time speed up the testing of changes. Mostly because the installation on its own takes ~4hrs plus whatever time it takes on the upgrade.

Patch-set: 36
CC: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>