Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > How do we verify this without first deleting the catalog index mirror?
> 
> We can validate on the new releases.

I'm afraid opm does not push the image as expected. I did a test and confirm that it only creates locally the new image, by tagging it as requested, but is still required to push it to our local registry to have it available.

The log when the image is pruned shows that it is only tagged but never pushed:

  INFO[0010] [podman build --format docker -f index.Dockerfile837211369 -t localhost:55000/test_index:v4.9 .]  packages="[performance-addon-operator sriov-network-operator ptp-operator]"

Patch-set: 1
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1