Update patch set 6
Patch Set 6:

I was checking this CR and I saw the failure in the last job. 
It seems to be a typo. The failed task showed a 404 Not Found in the following URL: https://raw.githubusercontent.com/e-minguez/ose-tests/master/rbac/osetests-openshift/conformance.yaml. If checking the redhat-tests role > tasks/setup.yml, variable "dci_openshift_agent_conformance" should be "openshift-conformance", but it is "openshift/conformance" (to change / for -)

Patch-set: 6