Update patch set 1
Patch Set 1:

Thanks for the examples, Pierre. I am also interested in this and I have some questions too. 
Is grpcurl intended to be used just for debugging purposes? Because the only change is to install grpcurl, but it is not used in any task. Or is it used indirectly in specific tasks related to the container registry?

Patch-set: 1
CC: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>