Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> (1 comment)
> 
> Hey Charles,
> 
> You probably saw the error already, I'm getting familiar with your change, I think all looks good, I just added some notes about the last test failure. I hope that helps, and I'll wait for your comments or new patch set. Thanks.

Thank you for help Manuel! I was planing to do some local tests to find the syntax I did not know the thing with variable with the name.
I have also remove the legacy way of testing as it was not working and for now, we want to have this working for one specific version at least, so let's try with 4.7 first

Patch-set: 9