Update patch set 1
Patch Set 1:

I see there's another change that takes care of this, so I'm unscheduling the checked job in the queue for this change as is not valid any longer.

Not sure if you want to have the changes from 22665 in here or viceversa, in any case please avoid using references in the commit to Jiras. Thanks!

Patch-set: 1
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>