Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 3:
> 
> hmmm the error obtained on libvirt is quite strange: "/bin/sh: oc: command not found", but the command syntax is correct. I will try to do a new recheck

The error is not in the script, the error is in a task, this change originally added a task in get-log.yml but then it was moved to the get-log bash script. The question is why is picking up a task that was modified and not the script?

Patch-set: 3
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1