Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> > Patch Set 3:
> > 
> > hmmm the error obtained on libvirt is quite strange: "/bin/sh: oc: command not found", but the command syntax is correct. I will try to do a new recheck
> 
> The error is not in the script, the error is in a task, this change originally added a task in get-log.yml but then it was moved to the get-log bash script. The question is why is picking up a task that was modified and not the script?

Found the error.

This commit (with Change-ID) is on top of the previous commit and not from master:

That's why the failure was on the previous commit but is "hidden" from this change:

```
$ git review -d 22666
$ git log --oneline -2
4061bc5 (HEAD -> review/user_samuel_elliott/cilab-273-cluster-network--output-in-get-logs) Obtain cluster network information in logs
20db932 Added a shell command to obtain the cluster's network configuration in YAML format, per Jira ticket CILAB-273.
```

Either the commit 20db932 has to be squashed from this change or a new clean change on top of master needs to happen. I let Sam decide which option he wants to do. I can help with the git part of it in either selection.

Patch-set: 3
Reviewer: Gerrit User 667 <667@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>