Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (3 comments)
> 
> > Patch Set 5:
> > 
> > (3 comments)
> > 
> > Thanks for checking this Ramon. It might be simpler to understand this with the results of a DCI Job, I haven't done tests in a few weeks, because I need to setup a special lab, but here is one from last week in Casa-Systems  https://www.distributed-ci.io/jobs/4a44c772-8a67-4faf-b1ac-0bdd4ca8578e/jobStates
> 
> Thanks for your response Manuel, I have understood what you commented (I've ACKed your messages too), that's what I figured out. 
> 
> However, I still have some questions, maybe from my ignorance 😂 if we have install, tests and teardown phases, shouldn't they be located into their corresponding phase (i.e. install tasks to be executed in d-o-a install phase, etc.)? Or is OCS following a different workflow and needs to execute everything sequentially?

Ramon, 
I see your point, when I started this, I was thinking to place OCS in a hook with each phase on the respective place, you know install, test, teardown, but I abandoned the idea in favor of integrating into d-o-a in one shot to keep it simple, as you can see by default only the deploy and test phases are executed, but I can omit the tests to avoid confusion changing the respective default variable, and I can leave the test and teardown options there, just in case someone ever needs them.

> By the way, in the job you shared, OCS code was executed twice: first one corresponds to the logic, but the second one starts after Launch partner install task (https://www.distributed-ci.io/jobs/4a44c772-8a67-4faf-b1ac-0bdd4ca8578e/jobStates#8a1ce89e-4c74-4bf5-9e15-130c2f50eae0:file268). Is this correct?

You are right,we should only apply once, but because this CR is not merged to d-o-a yet and I needed for Casa, I merged them manually in their installation, but then when I thought more carefully, I passed them as an install hook, so I ended up running it twice (because I didn't delete my initial manual merge), so yes I'm guilty but on the other hand, you can see how idempotent is the code :D 

Thanks for the observations!

Patch-set: 5