Update patch set 4
Patch Set 4:

I continue the discussion with Tony, here some comments to keep track of the conversation:
I confirmed we can create a SR-IOV network without the target namespace, we can add it on the network template, it simply won't create the object NetworkAttachmentDefinition because the target namespace does not exists.
If we don't create the NS and do not make any reference of it in the Network template, we can still create the network but we'll have to update it later when adding the NS, so maybe easier to just reference it from the beginning even if we don't create the NS.
I didn't extend my testing to make sure a pod using SR-IOV with this approach will actually work, but we'll leave the CI to test that :)
We also confirmed the example-cnf NS is not created in the example-cnf-config repo, but we are going to add it and remove it from this CR to test, so stay tune for the next patch!

Patch-set: 4