Update patch set 1
Patch Set 1:

The problem was that Task: "Get SriovNetworkNodeState for one of the worker node" was looking for interfaces in the state.resources[0].spec output, but since we are not provisioning the policies and network here anymore, that condition never happens and fails after 300 retries. Sending new patch.

Patch-set: 1
Work-in-progress: false