Update patch set 4
Patch Set 4:

Thanks Manuel for your comments. As Tony has also providing feedback, I would like to summarize a little bit what I have been observing in the jobs related to this change, also taking into account your comments and impressions:

-In fact, the results obtained in daily jobs and in this change are the same. If we compare this OCP-4.8 daily job (https://www.distributed-ci.io/jobs/ebc870d5-5c5f-40ad-8343-6b39eed55ab4/tests) with the passed OCP-4.8 job from this change (https://www.distributed-ci.io/jobs/bbe2929b-7bc6-4a75-ad98-a49fa42c351d/tests), the number of passed/skipped sriov tests is the same in both cases. 
And it also happens in OCP-4.10. If comparing a daily job (https://www.distributed-ci.io/jobs/ebca0194-e55a-4dff-a2d5-8912a2bd3441/tests) with the passed OCP-4.10 job from this change (https://www.distributed-ci.io/jobs/dc95cc03-9bda-4e1c-bbfc-fe9552ff2d4a/tests), we also have the same results in terms of sriov tests.
So, it is a "good" news, as we are having the same results than in the daily jobs, if we think on merging this change as-is now.

-I have also seen the rendering problem with the XML file uploaded, it is quite strange, in fact, but it seems that the tests whose name includes "strange" characters are not being rendered correctly. I guess this should be an issue to be reported in the openshift-kni/cnf-features-deploy

-Regarding sriov failed tests in OCP-4.10, I would bet openshift-kni/cnf-features-deploy repo has not been updated yet to include OCP-4.10. If you look in the Github repo (openshift-kni/cnf-features-deploy), there are branches for each OCP version until 4.9, or at least this is my impression. So, if it is true, I guess it is expectable to have these tests failed on OCP-4.10.

-Moreover, I would like to suggest something. For CNF Cert Suite, I save the output of the test suite execution in a log file (https://github.com/redhat-cip/dci-openshift-app-agent/blob/master/roles/cnf-cert/tasks/tests.yml#L73) and I upload it to DCI, in order to check the results. This has saved my life sometimes during debugging, and maybe it could be useful in this case too, to check the results obtained.

As a summary, I would say this change looks fine for OCP versions < 4.9. For OCP-4.10, maybe we can omit the tests until having the test suite release compatible with OCP-4.10. And also, it may be worth to contact that team to inform about the XML rendering problem, and also to include the output of the execution in a file to be submitted to DCI.

Patch-set: 4