Update patch set 14
Patch Set 14: Code-Review+2

At least, with the test I did here for dummy CR 22847 (https://www.distributed-ci.io/jobs/2eaf2989-410b-453c-aae6-02ce40cdee02/jobStates), it worked correctly (although the job failed, which was not related to dci-check-change). That is a change only related to d-o-a-a, and I executed dci-check-change by providing the CR and the kubeconfig path. In that way, no OCP cluster was installed and it directly executed d-o-a-a with the change (it can be seen that the debug message is present).
So, I think, it is fine to be merged.

Patch-set: 14
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2