Update patch set 1
Patch Set 1: Code-Review+2

Thanks Beto, this looks good. On a related topic this made me think about the default creds for SNO or OCP deployments using d-o-a. The oc-setup role performs the steps to include htpasswd auth and creates two users admin and nonadmin, which is cool but creds are not that difficult to break 😮. (just look at the d-o-a oc-setup role task roles/oc-setup/tasks/config.yml :D), maybe we should fix that in a diff PR or a partner can get an intrusion if they are not protecting the access to the API at the network level. Anyways it was just a comment. If you think we should throw a note about default creds of SNO deployments using DCI let me know and we can discuss.

Patch-set: 1
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2