Update patch set 34
Patch Set 34:

> Patch Set 34:
> 
> > Patch Set 34:
> > 
> > If the version is GA we have to know if operators are missing.
> > 
> > The GA version must work without any tweaks.
> 
> Then I believe there may be a conflict with pre-run task "set build:ga when build:candidate" (https://github.com/redhat-cip/dci-openshift-agent/blob/1a38d0e4af2d50558934813331d9e5040ac7b696/plays/pre-run.yml#L55).
> 
> Again, I'm not sure of this, but it looks like the operators may be missing if the release is a candidate, right? But the task above will cause the agent to treat candidate releases as ga, which may be the cause for the current failure.

What Pierre mentioned is correct, a candidate build should be considered GA build, because a candidate will only receive bug fixes and eventually becomes GA.

Additionally the changes not only has to do with operators, but also installation in itself.

Looking at the code helps to understand better this situation, rather than looking at the output of the job.

For example here it's explained that part of the considerations between GA and candidate:

https://github.com/redhat-cip/dci-openshift-agent/blob/master/plays/pre-run.yml#L51-L58

The initial reason to do build ga when is a candidate is due to a path in the URL of the installer, here you can see that (for old components) the URL is built based on the build:

https://github.com/redhat-cip/dci-openshift-agent/blob/master/roles/mirror-ocp-release/tasks/fetch.yml#L30

But we also took advantage of the build to determine whether a particular version of operators should be used or not:

https://github.com/redhat-cip/dci-openshift-agent/blob/master/roles/operators-mirror/tasks/mirror.yml#L2-L12

I understand that you mentioned that the change in the prerun "hides" the build later in the run, but for the operators we shouldn't care if a build is GA or candidate, it should be the same for both.  The only case we care is for nightly build, called dev.

hth

Patch-set: 34