Update patch set 6
Patch Set 6: Code-Review+2

(2 comments)

thx for the response Fred. In fact, I was thinking and the CR for fixing the tnf_test_example is not strictly required for this test, as you are passing it as argument (and it works), so I will see this can be merged. As I have a pending work of reviewing deeply dci-check-change, I will comment about possible improvements when doing that.

Patch-set: 6
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2