Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Change looks good, but I would like to see a job with OCP version higher than 4.5 to confirm everything is fine, just in case.
> By the way, dci-ci-bot marked some jobs as SUCCESS but they failed instead, is it a bug related to DCI to be reported?

Hello Ramon, 

Thanks for the review. Yeah, I already tested it on 4.7: https://www.distributed-ci.io/jobs/cbc613ad-cfd7-4e19-bf4f-dc53f6fc45df/jobStates. I am still not sure about why DCI is marking the job as failed and the bot as successful, my theory is that it is related to the the specific tags that are been executed. May need post-run to see if that fix the issue.

Patch-set: 3