Pull oc client for tests play to validate version
The tests.yml play performs a validation of the ocp
version, however in cases like upgrades where the
configure-jumpbox.yml play is executed before, this
cleans the directory where the oc client is stored
causing tests.yml play to fail. This patch applies
the tasks from oc-setup role to get the oc binary.

Change-Id: I15f862d0aa117865905a83b8adb3628776f6c391