Update patch set 3
Patch Set 3:

I touched based with Tony last week and we confirmed the issues observed in d-o-a-a are separate. 
On the other hand, if you do not mind I'm going to merge this, the last job shows all main phases green, and the upgrade succeeded, but there is a failed task in the bigip setup, not related to this patch.

Patch-set: 3