Update patch set 12
Patch Set 12:

I have a new question here, and also applies to CR 23045. I checked the last jobs on libvirt, and I saw that the d-o-a component appears twice in the d-o-a job (one comes from the git repo, and the other one from the rpm), and that also the d-o-a-a component is created. And the same happens for d-o-a-a in the d-o-a-a job but in the opposite way. Do we want to have repeated components in the job? And also, do we also want to have the d-o-a-a component in the d-o-a job (and the opposite in d-o-a-a job)?

Patch-set: 12
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>