Update patch set 7
Patch Set 7:

Hi Nacho! As commented in today's daily, for this change: https://softwarefactory-project.io/r/c/dci-openshift-app-agent/+/23080, it is needed to install Ansible's podman collection prior to running these tasks. So, is it possible to include the installation of that collection in this task? It is as simple as running the following command:

```
ansible-galaxy collection install containers.podman
```

Patch-set: 7
CC: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>