Update patch set 4
Patch Set 4: Code-Review+1

> Patch Set 2: Code-Review+1
> 
> I've never undertood why we did this 😄 as the operator and mc takes care of those steps.  In any case I'd only add a small delay for the mc to be created and then just have the checknodes do it's validation.

I thought getting the validation tasks to get the PAO profile MC and the respective rendered-* were just to be sure it was created, but I agree to remove this as it seems it's a very small delay and the the tasks in nodes-ok.yml will take care. I've seen other configurations take longer to trigger the MC, but PAO is not the case.

Patch-set: 4
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1