Update patch set 1
Patch Set 1:

(1 comment)

Answering Nacho's comment. And apart from this, I think this CR is needed in fact, because for example in this SPK job [1], the get-logs.yml is failing because it cannot find dciclient module, and I guess it is because it is missing the delegate_to clause.

[1] https://www.distributed-ci.io/jobs/69820e4d-d9e7-49ea-9452-53ba9dad8215/jobStates#84e13b83-f781-418b-bdf4-83ddfb7c5baa:file7

Patch-set: 1
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>