Add custom number of mcp nodes for prepare-cnf
By default the number of the mcp member is equal to the worker nodes.

This change allows to specify the number of nodes expected
in the workers MCP with variable "mcp_worker_nodes"

If we want to have a custom number of nodes on our mcp with the
option customize_extramanifests_path, we need to be able to set this
number.

Change-Id: I12e22cdf9975b5ea71d4ab27fce59466a5f8dc82