Update patch set 2
Patch Set 2: Code-Review+1

Pierre, the change looks good to me, but I suggest to update the commit message, something like: this change allows to specify the number of nodes expected in the workers MCP with variable "mcp_worker_nodes". If we use customized_extramanifests_path to include a new MCP, that is secondary. wdyt?

IMHO, I do not think it's necessary to document it, the table in the README contains mustly dci_* vars, so if you want to document it maybe the inventory is a better place.

Patch-set: 2
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1