Update patch set 3
Patch Set 3:

> I see your point, I will adapt it :)
Thanks, it makes more sense for me now! 

> The inventory is documented in the README, isn't it ?
It is somehow documented, but as you might have seen, it's just an example of an inventory to use, and we keep that example with the minimum required variables to deploy, I was mostly thinking to add and document this variable in the respective inventory file you use for the deployment. However to keep it in the DCI repo, maybe add a small README or a vars/main.yml in the roles/prepare-cnf/ directory with a brief description of the variable? 

Personally I wouldn't worry to document this var, there is a bunch more not documented either, but if we want to start doing this as a good practice, maybe the role directory where the tasks using this variable are located could be a good start.

Patch-set: 3
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>