Update patch set 28
Patch Set 28:

Ramon, thanks for providing the f5 references and for preparing this change.

I have a few questions/comments:

I read some comments in this PR from Tony (Aug 20) about using separate PAO profiles, and use variable "performance_definition" to define which profile to use, May I bring this question back? Just trying to understand if might be better to define a profile for each type of configuration we need (specially if f5 spk requires specific settings), or are we trying to produce a profile that could match all scenarios? 

About the CPU configuration, I share Jorge's opinion, so let's sort this out together: 

In Dallas servers we have 2 sockets => 20 cores per socket => each socket has 2 threads because of HT, so that give us 80 cores. 

Per the f5 doc there are some limitations[1]. Also reading about the manual-cpu guidelines[2] makes me think we should at least consider the following:

- Should we disable HT, (we don't need to disable in the BIOS, we can control during runtime[3], we can discuss this separately) 
- We should not use core IDs 0 and 1 per socket.
- Assuming we disable HT, then socket 1 should contain cores: 0-19, and socket 2: 20-39,

Finally, I was checking patchset 26/27 when you were testing two groups of hugepages configuration, and I noticed we are using (defaultHugepagesSize: "1G"): and Per Beto's comments in Aug 30: "It is possible but it is important to notice that setting hugepages default size to 1G will remove all 2M related folders from the node and it will be impossible to configure 2M hugepages under the node." So we might want to remove that default setting if we want to setup multiple hugepages configurations.

I think our current PAO basic profile somehow provides a viable configuration for most workloads, but if we want to be strict and follow the recommendation of f5 or other workload we need to shape it a little bit more. For example I was thinking if we prepare an f5 PAO profile we can be even more accurate and use the cores of the NUMA node where f5 interfaces will be connected only ( ens3f0 and ens3f1 in NUMA node 1), just some thoughts. 


[1]https://clouddocs.f5.com/service-proxy/latest/spk-openshift-requirements.html#scheduler-limitations
[2]https://clouddocs.f5.com/service-proxy/latest/spk-manual-cpu.html#allocation-guidelines
[3]https://access.redhat.com/solutions/rhel-smt

Patch-set: 28