Update patch set 29
Patch Set 29:

The change is working fine in cluster5 and 6. Last job failed because example-cnf failed in a step that is not related to this change, but the loadbalancer pod (the one affected by the performance profile) was installed correctly, so it is working. This change is ready to be merged, then we can think afterwards on refactoring the code to simplify the definition of performance profiles

Patch-set: 29