Update patch set 29
Patch Set 29: Workflow+1

Despite the failures on the previous jobs, the purpose of this change, which is to include 2M hugepages, is working fine in all cases, and example-cnf is being deployed correctly too. So, after discussing this with Manuel and Charles, we have seen that this CR can be merged. The update of VCP pipeline will be handled in the corresponding PR on github.

Patch-set: 29
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Workflow=+1