Update patch set 47
Patch Set 47:

> Patch Set 47:
> 
> > Patch Set 47:
> > 
> > Hi Nacho. Just to add some comments, because I have been checking your changes and I have seen, at least in the last job you executed, that it is working fine and that it saves some minutes in the execution of the upgrade tasks changed here. However, I wanted to comment something:
> > 
> > -I was wondering what is the real improvement we want to have with this change. At least, it is not clear for me. For example, we are currently running the OCP upgrade pipeline in the daily jobs, from 4.7.24 to 4.8.21, and they are working. Is it something more apart from, maybe, save time during execution or using more complex commands/k8s module?
> > -In Monitor upgrade until complete task, no logs are shown in the tasks. Could it be possible to show the output of the shell commands, or to generate a .log file and upload it to DCI to analyze it?
> 
> Hi Ramón, 
> The tricky part with this change is that is addresses the worst case scenario for the upgrade process, which is that the upgrade fails.
> The original code either waited for the upgrade to successfully complete or for the 2h timeout to expire. This meant early errors wouldn't be detected and the job/pipeline would be kept blocked for 2 hours.
> In its current version, this change attempts to identify an upgrade failure if a failure condition remains active for 15 minutes so we could save up to 1h45 minutes in an upgrade that would eventually fail.

I think I got it, thanks Nacho! For the moment, I will wait for more jobs or possible updates of the code before giving a positive review, because my fear is that it has passed almost 2 months since the last patchset, and however, there are still a lot of failed jobs. Maybe you need to play a little bit with the timings you are giving for the async task or something like this to ensure the job passed in the worst case you commented?

And also, the fact related to obtain the logs generated during the Monitor upgrade until complete task, if possible.

Patch-set: 47
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>