Update patch set 9
Patch Set 9:

> Patch Set 8:
> 
> (1 comment)
> 
> Hi Beto, the code LGTM. I checked the last success job though, and the file csi-report.txt seems to be empty: https://www.distributed-ci.io/jobs/a269e13b-fca6-48be-9d26-410b0fc80499/files Don't know if it's supposed to be this way.

Hello Tatiana, 

That is weird, in that job I was playing a bit trying to use the tags inside the role but didn't work. Not sure why the file was empty if the tasks was not executed. 

The following are the 2 jobs where the test is executed and the file has the proper results. 

VCP
https://www.distributed-ci.io/jobs/94054bcc-c5c4-4507-8bdc-39f0920cab67/files

Vanilla
https://www.distributed-ci.io/jobs/1f663e6d-0775-4504-8aa5-4ae194a56a59/files

Will check about your other recommendation.

Cheers

Patch-set: 9