Update patch set 2
Patch Set 2:

I'm not really sure if what I am doing is correct, because it seems that delegate_to is not working fine when combining with include_tasks. If somebody has more ideas, it would be great, but if not, maybe this change will not be needed because, in the end, the current code works. However, it would be a mess if, for each task included in get-logs (for example), you need to put delegate_to due to the way the error and failure processes are designed.

Patch-set: 2