Update patch set 4
Patch Set 4:

Recreating the jobs again, the two last ones will be the ones running.
The first one, as that spk_version does not exist, it will fail in the install stage (which in theory is run in provisioner, but it should be delegated to localhost when running the teardown_failure), and the second one should fail in the test stage (teardown will be run in localhost).

Patch-set: 4