Update patch set 4
Patch Set 4: Code-Review+1

Fortunately, in the two previous failed jobs, we are able to see one case (the first one) in which it fails in the test stage, which is executed in localhost, and another case (the second one) in which it fails in the install stage, which is executed in provisioner (but the failure process is delegated to localhost). Both failure processes worked correctly and without failures, so this is ready to be merged.

Patch-set: 4
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1