Update patch set 1
Patch Set 1:

Just to understand this. If we are already waiting for having all the nodes in Done state, checking the MC status, do we really need to wait a minute before these tasks? Is this to avoid failing in the "Wait for ALL nodes to be in a Done state" task because the MC is not applied? In that case, may the solution be to increase the number of retries to cover this minute that is being added in this change?

Patch-set: 1
CC: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>