Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Just to understand this. If we are already waiting for having all the nodes in Done state, checking the MC status, do we really need to wait a minute before these tasks? Is this to avoid failing in the "Wait for ALL nodes to be in a Done state" task because the MC is not applied? In that case, may the solution be to increase the number of retries to cover this minute that is being added in this change?

The reason of the delay prior checking the status is because the mc may take some time to render.
For example, if something triggers a new mc and we start checking the status of the nodes immediately after, we won't catch the new rendered configuration, thus the nodes-ok check will pass because all the nodes are in a "Done" state.  That's the reason of the delay. We're doing the same in other places, for example in example-cnf-config (https://github.com/dci-labs/example-cnf-config/blob/master/testpmd/hooks/roles/mirror-rh-nfv/tasks/main.yml#L47-L49)

hth

Patch-set: 1